Major change: replaced modals with pages
Removed the modals and single page application structure as it was causing usability issues and frustration to users. Replaced it with html pages, added Navbar on top of pages for now to allow users to navigate back as needed in addition to the natural back navigation through the browser button that users are familiar with.